Add files via upload
Attached is the final project for my STAT601 - Applied Statistics course. The goal of this project was to create a PP presentation that unpacked password typing speed change over time. We were given a dataset containing the typing speed of a few hundred subjects over 400 trials. Given just a .xls file our goal was to create a model to best fit the password typing speed of the average user over time.

Attached is two files, the PowerPoint presentation with my final results and outputs as well as the source code.